extract Index2D code to common (+ trivial test)
