+++ test/unit/CMakeLists.txt
@@ -11,5 +11,6 @@
 DLAF_addTest(test_util_math SOURCES test_util_math.cpp USE_GTEST_MAIN)
 DLAF_addTest(test_tile SOURCES test_tile.cpp LIBRARIES PRIVATE DLAF_gtest_hpx_main)
 
+add_subdirectory(common)
 add_subdirectory(memory)
 add_subdirectory(communicator)
